pink-s3: Fix compilation error
