+++ crates/pink-libs/s3/Cargo.toml
@@ -3,7 +3,7 @@ description = "Simple S3 client for Phala's pink"
 homepage = "https://github.com/Phala-Network/phala-blockchain"
 license = "Apache-2.0"
 name = "pink-s3"
-version = "0.4.0"
+version = "0.4.1"
 edition = "2021"
 keywords = ["fat-contract", "pink", "ink", "S3"]
