+++ crates/pink-libs/s3/example/Cargo.toml
@@ -1,0 +1,28 @@
+[workspace]
+[package]
+name = "example"
+version = "0.1.0"
+authors = ["<author>"]
+edition = "2021"
+
+[dependencies]
+ink = { version = "4", default-features = false }
+pink-s3 = { version = "0.4.0", default-features = false, path = ".." }
+scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
+scale-info = { version = "2.3", default-features = false, features = ["derive"], optional = true }
+
+
+[dev-dependencies]
+pink-extension-runtime = { path = "../../../pink/pink-extension-runtime" }
+
+[lib]
+path = "lib.rs"
+
+[features]
+default = ["std"]
+std = [
+    "pink-s3/std",
+    "ink/std",
+    "scale-info/std",
+]
+ink-as-dependency = []
